requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.0
daphne==4.0.0
Django==4.2.3
django-bootstrap5==23.3
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
gTTS==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
multidict==6.0.4
openai==0.27.7
packaging==23.2
Pillow==10.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
zope.interface==6.1
